requirements.txt
numpy
Flask
gunicorn
scikit-learn
pandas
PyYAML
evidently==0.2.0
dill
matplotlib

sklearn-pandas 
imblearn
kneed
